Update phenotype criteria handler and test.
